feature: AI 요약 배경 이미지 추가
